requirements.txt
django==5.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.10
whitenoise==6.8.0
django-cors-headers==4.6.0
django-rest-framework
djangorestframework-simplejwt
django-phonenumber-field 
phonenumbers
dj-database-url==3.0.1
django-filter==25.1
pillow==12.1.0
firebase-admin